+++ src/domain/ledger/index.ts
@@ -1,2 +1,3 @@
 export * from './entities';
 export * from './errors';
+export * from './service';